feat:rajout de createConversation
